Modified denominators from partial derivative and the sign from the first derivative value after validating force-computing routine
Two partial derivatives looks like divided by cv_threshold and
distance_cut value and a sign in the partial value was fixed. Further
changes are denoted in the comments - Finally I’ve compared to the
original MS-UCG 2 code and it worked exactly same as the original one
for three particle cases, 1125 particles in the slab system for very
first 100 steps.